feat(posts): basic access control list for posts

+++ post-service/src/app/commands/post/delete/delete.command-handler.ts
@@ -1,5 +1,6 @@
+import { PostAccessControlList } from '@/app/boundaries/acl';
 import { DeletePostOutput } from '@/app/boundaries/dto/output';
-import { NotFoundException } from '@/app/boundaries/errors';
+import { ForbiddenException, NotFoundException } from '@/app/boundaries/errors';
 import { PostRepository } from '@/app/boundaries/repository';
 import { Command } from '@/app/commands/common';
 import { Inject } from '@nestjs/common';
@@ -11,6 +12,9 @@ export class DeletePostCommandHandler extends Command<
   DeletePostCommand,
   DeletePostOutput
 > {
+  @Inject(PostAccessControlList)
+  private readonly _postACL: PostAccessControlList;
+
   @Inject(PostRepository)
   private readonly _postRepository: PostRepository;
 
@@ -18,6 +22,13 @@ export class DeletePostCommandHandler extends Command<
     const post = await this._postRepository.getById(input.postId);
     if (!post) {
       throw new NotFoundException('Post does not exist');
+    }
+
+    const hasPermission = await this._postACL.canDelete(post.owner, post);
+    if (!hasPermission) {
+      throw new ForbiddenException(
+        'You do not have permission to delete this post',
+      );
     }
 
     await this._postRepository.delete(post.id);
